backend/pyproject.toml
[project]
name = "restaurant-llm-chat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0",
  "redis>=5.0.7",
  "aiokafka>=0.10.0",
  "orjson>=3.10.6",
  # LLM orchestration
  "langchain>=0.2.12",
  "langgraph>=0.2.34",
  "langsmith>=0.1.99",
  "crewai>=0.51.0",
  "pyautogen>=0.3.2",
]
